Move redux files to their own folders.
