+++ src/redux/configureStore.js
@@ -1,8 +1,8 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import thunk from 'redux-thunk';
-import missionsReducer from './missions';
-import rocketsReducer from './rockets';
+import missionsReducer from './missions/missions';
+import rocketsReducer from './rockets/rockets';
 
 const reducers = combineReducers({ missionsReducer, rocketsReducer });
